# packages/backend/src/swagger.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/api/marketplace/apps/search": {
      "get": {
        "operationId": "searchApps",
        "parameters": [
          {
            "name": "category",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "ai",
                "automation",
                "books",
                "data",
                "development",
                "featured",
                "finance",
                "gaming",
                "media",
                "music",
                "network",
                "photography",
                "security",
                "social",
                "utilities"
              ]
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "string",
                  "pattern": "^(?:(?!^-0$)-?(?:(?:0|[1-9]\\d*)))$"
                }
              ],
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "storeId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchAppsDto"
                }
              }
            }
          }
        },
        "tags": [
          "Marketplace"
        ]
      }
    },
    "/api/marketplace/apps/{urn}/image": {
      "get": {
        "operationId": "getImage",
        "parameters": [
          {
            "name": "urn",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Marketplace"
        ]
      }
    },
...
    "/api/marketplace/create": {
      "post": {
        "operationId": "createAppStore",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAppStoreBodyDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppStoreDto"
                }
              }
            }
          }
        },
        "tags": [
          "Marketplace"
        ]
      }
    },
...
    "/api/marketplace/{id}": {
      "patch": {
        "operationId": "updateAppStore",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAppStoreBodyDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateAppStoreDto"
                }
              }
            }
          }
        },
        "tags": [
          "Marketplace"
        ]
      },
      "delete": {
        "operationId": "deleteAppStore",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Marketplace"
        ]
      }
    },
...
    "/api/user-config/{urn}": {
      "get": {
        "operationId": "getUserConfig",
        "parameters": [
          {
            "name": "urn",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUserConfigDto"
                }
              }
            }
          }
        },
        "summary": "Get the user configuration for an app",
        "tags": [
          "User Config"
        ]
      },
      "put": {
        "operationId": "updateUserConfig",
        "parameters": [
          {
            "name": "urn",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserConfigDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update the user configuration for an app",
        "tags": [
          "User Config"
        ]
      }
    },
    "/api/user-config/{urn}/enable": {
      "post": {
        "operationId": "enableUserConfig",
        "parameters": [
          {
            "name": "urn",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Enable the user configuration for an app",
        "tags": [
          "User Config"
        ]
      }
    },
    "/api/user-config/{urn}/disable": {
      "post": {
        "operationId": "disableUserConfig",
        "parameters": [
          {
            "name": "urn",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Disable the user configuration for an app",
        "tags": [
          "User Config"
        ]
      }
    },
...
  },
  "info": {
    "title": "Runtipi API",
    "description": "API specs for Runtipi",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "SearchAppsDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "available": {
                  "type": "boolean"
                },
                "categories": {
                  "type": "array",
                  "items": {
                    "enum": [
                      "ai",
                      "automation",
                      "books",
                      "data",
                      "development",
                      "featured",
                      "finance",
                      "gaming",
                      "media",
                      "music",
                      "network",
                      "photography",
                      "security",
                      "social",
                      "utilities"
                    ]
                  }
                },
                "created_at": {
                  "type": "integer",
                  "minimum": 0
                },
                "deprecated": {
                  "type": "boolean"
                },
                "id": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "short_desc": {
                  "type": "string"
                },
                "urn": {
                  "type": "string"
                },
                "supported_architectures": {
                  "type": "array",
                  "items": {
                    "enum": [
                      "amd64",
                      "arm64"
                    ]
                  }
                }
              },
              "required": [
                "available",
                "categories",
                "created_at",
                "deprecated",
                "id",
                "name",
                "short_desc",
                "urn"
              ]
            }
          },
          "total": {
            "type": "number"
          },
          "nextCursor": {
            "anyOf": [
              {
                "type": "string"
              }
            ]
          }
        },
        "required": [
          "data",
          "total"
        ]
      },
...
      "CreateAppStoreBodyDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 16,
            "minLength": 1
          },
          "url": {
            "type": "string",
            "format": "uri"
          }
        },
        "required": [
          "name",
          "url"
        ]
      },
      "AppStoreDto": {
        "type": "object",
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "url": {
            "type": "string"
          }
        },
        "required": [
          "enabled",
          "name",
          "slug",
          "url"
        ]
      },
...
      "UpdateAppStoreBodyDto": {
        "type": "object",
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "enabled",
          "name"
        ]
      },
      "UpdateAppStoreDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ]
      },
...
      "GetUserConfigDto": {
        "type": "object",
        "properties": {
          "appEnv": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "dockerCompose": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "isEnabled": {
            "type": "boolean"
          }
        },
        "required": [
          "appEnv",
          "dockerCompose",
          "isEnabled"
        ]
      },
      "UpdateUserConfigDto": {
        "type": "object",
        "properties": {
          "appEnv": {
            "type": "string"
          },
          "dockerCompose": {
            "type": "string"
          }
        },
        "required": [
          "appEnv",
          "dockerCompose"
        ]
      }
    }
  }
}